Replace 'entity' identifier example from cdl.md (#1982)
In the "Keywords & Identifiers" section there's an example of the
keyword `entity` being used as an identifier, which seems confusing and
"asking for trouble" :-)

Moreover, a little further on, in a "Warning" section, it states
"Delimited identifiers in general, but in particular non-ASCII
characters, _or keywords as identifiers_ should be avoided as much as
possible, for reasons of interoperability." (emphasis mine).

In addition, the definition of the `entity` aspect also had an `Integer`
key called `ID`, which is perhaps not to be universally encouraged
(better to use `cuid` with its `UUID` type).

So I reworked this section.

Co-authored-by: Daniel Hutzel <daniel.hutzel@sap.com>