audit logging: enforce data integrity for data subject id fields (#2116)
A [recently opened issue in the audit log plugin
repo](https://github.com/cap-js/audit-logging/issues/180) showcased an
error scenario, where the stakeholder tried to model an entity
(`@PersonalData.EntitySemantics: 'Other'`) that references a data
subject by way of `association to one DataSubject
@PersonalData.FieldSemantics: 'DataSubjectID'`. In this scenario it is
possible to produce an error response from the audit logging service, if
no value for `data_subject.id` is present. However, in the context of
audit logging, IMO this should never be possible. To unilaterally
'solve' the problem we might want to update our docs to specify that
fields that are supposed to hold the data subject id must not be null.

---------

Co-authored-by: sjvans <30337871+sjvans@users.noreply.github.com>
Co-authored-by: René Jeglinsky <rene.jeglinsky@sap.com>