highlight package freeze before deployment (#2158)
The tip is quite easy to overlook because it does not look like an
actionable item. The `package-lock.json` needs to be updated if there
already is one... which is not uncommon, so I think we should include it
as a regular step.

After a `cds add mta` we recommend to:
```
run npm update --package-lock-only to freeze dependencies
```

So I'd suggest we also document it here instead of the `npm update`. The
`npm update` also changes all other versions which were not changed in
the `package.json`.
Of course it is important to regularly run `npm update` to use the
latest versions.