Align with entity naming best practice (#2201)
See:
https://cap.cloud.sap/docs/guides/domain-modeling#naming-conventions
There are plenty of entity names in examples in this CDL topic. Some of
them are "real" in that they are genuine representations such as Orders
and Products. These should be subject to the naming convention best
practice of being pluralised.

Others are clearly illustrative only, such as the Foo, Bar, Baz set and
those conveying meaning for the given example such as EmployeeView,
SomeView, P, Proj, SomeExposedEntity and so on. These should not be
subject to the naming convention.

Note: Perhaps we should add something to the naming convention section
itself to qualify the boundary of where it applies, i.e. "mostly" at the
schema level, less so at the service definition level?